face_recongnition拼错了
现在接收一个灰度图片的np数组就可以预测情绪了
上传了demo
train是训练模型的

+++ face_emotion_recongnition/utils/image/resize_image.py
@@ -22,4 +22,4 @@ def resize_image(image, size, method=BILINEAR):
         )
     
     else:
-        return cv2.resize(image, size, interpolation=method)
+        return cv2.resize(image, size, interpolation=method).astype(int)
